\name{infer-package}
\alias{infer-package}
\alias{infer}
\docType{package}
\title{
InFER: Inference For Epidemic-related Risk
}
\description{
This package contains a set of tools for performing inference on epidemic models.
}
\details{
\tabular{ll}{
Package: \tab infer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-02\cr
License: \tab GNU Public License version 3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Chris Jewell (c) 2012

Maintainer: Chris Jewell <c.p.jewell@massey.ac.nz>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
